Refactor the following: - Combine cli.py & v1.py - Rename scripts/ to cli/ - Rename Model.items_iter to Model.iterate - Moved item_asset_types under api/ since it is not click-specific
