+++ planet/api/__version__.py
@@ -1,1 +1,1 @@
-__version__ = '1.4.3'
+__version__ = '1.4.4'
